Better Server Shutdown and Logger Integration (#15)
* refactor: allow changing the number of cached pages

* refactor: add `from_u8` function

- Add a `TryInto` for `&[u8]`

* refactor: modify pager to track pages in memory by default

`Pager` now tracks all pages in memory and requires
users to call `flush` to clear up and persist the memory.

* refactor: logger ensures in-memory pager is always up to date

Logger keeps track of an up-to-date pager for in-memory
modification. It only permanently writes out when a GlobalCheckpoint
is logged. During startup it'll re-read previous state before
the server crashed and re-apply it.

* refactor: only compact log when server is dropped

* refactor: allow users to shutdown server with Ctrl-C

* chore: remove unneeded imports